gcc.dg/pr90838-2.c: Replace long with long long
Since gcc.dg/pr90838-2.c is only for 64-bit integer, replace long with
long long for ILP32 targets.

	* gcc.dg/pr90838-2.c (ctz4): Replace long with long long.

+++ gcc/testsuite/gcc.dg/pr90838-2.c
@@ -26,9 +26,9 @@ static const char table[128] = {
     57, 58, 59, 60, 61, 62, 63, 64
 };
 
-int ctz4 (unsigned long x)
+int ctz4 (unsigned long long x)
 {
-  unsigned long lsb = x & -x;
+  unsigned long long lsb = x & -x;
   return table[(lsb * magic) >> 58];
 }
